add key-in amount edittext and effect other items

+++ app/src/main/java/com/rakuten/assignment/bean/CountryExchangeRate.kt
@@ -1,12 +1,10 @@
 package com.rakuten.assignment.bean
 
-import java.math.BigDecimal
-
 data class CountryExchangeRate(
     val iso: String,
     val rate: Double,
-    val amount: BigDecimal,
-    val base : String
-){
-    val baseRate : Double = 1 / rate
+    val amount: Double,
+    val base: String
+) {
+    val baseRate: Double = 1 / rate
 }
